+++ src/main/java/labspanish/utilidades/DistribucionArribo.java
@@ -11,7 +11,7 @@ package labspanish.utilidades;
 public class DistribucionArribo implements Distribucion{
 
     @Override
-    public double getProbabilidad(double event) {
+    public double getProbabilidad(int event) {
         if(event == 10){
             return 0.35;
         }
@@ -27,7 +27,7 @@ public class DistribucionArribo implements Distribucion{
 
 
     @Override
-    public double getEvento(double acumulativo) {
+    public int getEvento(double acumulativo) {
         if(acumulativo < 0.35){
             return 10;
         }
